+Memories routes

+++ src/routes/memories.ts
@@ -3,18 +3,16 @@ import { prisma } from '../lib/prisma'
 import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  // app.get('/users', async () => {
-  //   const users = await prisma.user.findMany({
-  //     select: {
-  //       name: true,
-  //     },
-  //   })
-  //   return users
-  // })
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
 
   // Get all memories
-  app.get('/memories', async () => {
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -27,7 +25,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
 
   // Get memory by id
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -38,6 +36,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub)
+      return reply.status(401).send()
+
     return memory
   })
 
@@ -56,7 +58,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '3007f88e-fb2e-4178-b616-3a73311bc811',
+        userId: request.user.sub,
       },
     })
 
@@ -64,7 +66,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
 
   // Update memory
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -77,7 +79,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const { id } = paramsSchema.parse(request.params)
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) return reply.status(401).send()
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -92,11 +102,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
 
   // Remove memory
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) return reply.status(401).send()
 
     await prisma.memory.delete({
       where: {
